Refactor exercises and its nested end points to use the new routing style
