+++ app/controllers/api/v8/apidocs_controller.rb
@@ -93,9 +93,15 @@ class Api::V8::ApidocsController < ActionController::Base
     Api::V8::CoursesController,
     Api::V8::Courses::PointsController,
     Api::V8::Courses::Users::PointsController,
+    Api::V8::Courses::ExercisesController,
+    Api::V8::Courses::Exercises::PointsController,
+    Api::V8::Courses::Exercises::Users::PointsController,
     Api::V8::Organizations::CoursesController,
     Api::V8::Organizations::Courses::PointsController,
     Api::V8::Organizations::Courses::Users::PointsController,
+    Api::V8::Organizations::Courses::ExercisesController,
+    Api::V8::Organizations::Courses::Exercises::PointsController,
+    Api::V8::Organizations::Courses::Exercises::Users::PointsController,
     Api::V8::ExercisesController,
     Api::V8::SubmissionsController,
     Api::V8::UsersController,
